Add the end of the tagging process, make sure all legal positions not drawn have been processed
This actually made me tweak `outcome_from_captures_promotion` to also return information that all moves are capture/promotion or if there exist other kind of moves.
In the former case, this mean the position is "closed"/"irreversible" as a break in the material config is inevitable and that the position's outcome is only determined by those moves.

No change on the final table generated since `Reports` are erased only to keep the `Outcome` when writing to disk