Find the issue (unfixed) with generation + show how the diffs articulate between the new and the old table with `diff`
Found the root issue:
The algo starts from Mate positions then proceed backward.
Win(0) -> Win(1) -> ...
This work well when there are no capture/promotion, but it breaks with them.
For example Win(0) -> Win(1) -> Win(2) -> Win(1) -> Win(2)... because it finds a capture that lead to mate.

This is clearly wrong, and the current algo is dependent on the order in which positions are processed.

Because the new queue involved processing pos in a new order, the table changed but the **root issue** is the same as with the old queue.

This is shown by the fact that almost half positions have found a better solution, and the other regressed.

Normally (expect for other bugs), a good implemention would **only** improve results.

Found 5286 differences
Old is better: 2785 cases New is better: 2501

The solution would be to drop the queue system and work directly on the `all_pos` array of index (which would mean only reversible would be supported, and syzygy compression would only be possible afterwards).
That mean also potentially be able to remove the need for `Report`