Store the newly tagged positions on the all_pos vec directly
Allowing to get rid of `OneQueue` altogether. The `queue.rs` file is not removed yet due to a git issue.

The concept of pass does not exist anymore when looking for drawn positions, because the positions are modified during the pass, eg: position 111 can unmove to 112 and claim it is drawn too, and during the same pass, position 112 will be tagged.